back-end-app/apiupdateworker/requirements.txt
pika
yfinance
psycopg2-binary